overthinking too much

+++ hw6/hw6-pintos-starter/src/lib/user/stdlib.h
@@ -1,8 +1,9 @@
 #include <stdbool.h>
+
 #ifndef __LIB_USER_STDLIB_H
 #define __LIB_USER_STDLIB_H
 
-void* malloc (size_t size);
+void* malloc (int size);
 void free (void* ptr);
 void* calloc (size_t nmemb, size_t size);
 void* realloc (void* ptr, size_t size);
